+++ parts/header_file/files/_files/string/string_data_t/null_string_data_t.hpp
@@ -37,7 +37,7 @@ struct null_string_data_t final:base_string_data_t<char_T>,instance_struct<null_
 	[[nodiscard]]virtual ptr_t do_pop_back([[maybe_unused]]size_t size,[[maybe_unused]]ptr_t& self)override final{ return this; }
 	[[nodiscard]]virtual ptr_t do_pop_front([[maybe_unused]]size_t size,[[maybe_unused]]ptr_t& self)override final{ return this; }
 
-	[[nodiscard]]virtual size_t get_memory_cost()override final{return 0;}
+	[[nodiscard]]virtual double get_memory_cost()override final{return 0;}
 };
 template<typename char_T>
 inline null_string_data_t<char_T> null_string_data{};
